Implement generic table hooks
