+++ client/packages/common/src/ui/layout/tables/index.ts
@@ -28,6 +28,7 @@ export * from './columns';
 export * from './utils';
 export * from './context';
 export * from './components';
+export * from './material-react-table';
 export const DEFAULT_PAGE_SIZE = 25;
 
 export const placeholderRowStyle: AppSxProp = {
