+++ client/packages/common/src/ui/layout/tables/material-react-table/commonProps.ts
@@ -1,0 +1,5 @@
+import { MRT_TableOptions } from 'material-react-table';
+
+export const commonProps: Partial<MRT_TableOptions<any>> = {
+  //
+};
