+++ client/packages/common/src/ui/layout/tables/material-react-table/index.ts
@@ -1,0 +1,3 @@
+export * from './MaterialReactTable';
+export * from './usePaginatedMaterialTable';
+export * from './useNonPaginatedMaterialTable';
